fix: put everything together (through the navbar)
