+++ src/app/layout.tsx
@@ -1,19 +1,20 @@
 import { ClerkProvider } from "@clerk/nextjs";
 import "./globals.css";
-import { Patua_One } from "next/font/google";
+import { Lora } from "next/font/google";
 import "./globals.css"; // Ensure your global styles are imported
 
 // 1. Configure the Patua One font
-const patua = Patua_One({
-  weight: "400", // Patua One only has a 400 weight available
+const lora = Lora({
   subsets: ["latin"],
-  variable: "--font-patua", // This creates the hook for Tailwind
+  weight: ["400", "500", "600", "700"],
+  variable: "--font-lora",
 });
+
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <ClerkProvider>
       <html lang="en">
-        <body className={`${patua.variable} antialiased`}>{children}</body>
+        <body className={`${lora.variable} antialiased`}>{children}</body>
       </html>
     </ClerkProvider>
   );
