+++ src/components/EventDetails.tsx
@@ -70,8 +70,8 @@ const EventDetails: React.FC<EventDetailsProps> = ({ eventData }) => {
       border: "1px solid #527ABE",
       borderRadius: "5px",
       padding: "10px 15px 10px 15px",
-      maxWidth: "45%",
-      maxHeight: "35%",
+      maxWidth: "100%",
+      maxHeight: "auto",
       display: "flex",
       flexDirection: "column",
       gap: "25px",
