+++ src/components/Navbar.tsx
@@ -1,10 +1,12 @@
 "use client";
 
 import Link from "next/link";
+import Image from "next/image";
 import { usePathname } from "next/navigation";
 import styles from "@/styles/Navbar.module.css";
 
 type NavMode = "Admin" | "VolunteerNotLoggedIn" | "VolunteerLoggedIn";
+
 type NavItem = {
   label: string;
   href: string;
@@ -18,12 +20,10 @@ const NAV_LINKS: Record<NavMode, NavItem[]> = {
     { label: "Report", href: "/report" },
     { label: "Account", href: "/account" },
   ],
-
   VolunteerNotLoggedIn: [
     { label: "Calendar", href: "/calendar" },
     { label: "Login", href: "/login" },
   ],
-
   VolunteerLoggedIn: [
     { label: "Calendar", href: "/calendar" },
     { label: "My Events", href: "/events" },
@@ -36,10 +36,19 @@ const isActivePath = (pathname: string, href: string) => pathname === href || pa
 
 export default function NavBar({ mode }: { mode: NavMode }) {
   const pathname = usePathname();
-  const links = NAV_LINKS[mode];
+  const links = NAV_LINKS[mode] ?? [];
 
   return (
     <nav className={styles.nav}>
+      {/* LEFT SIDE */}
+      <div className={styles.brand}>
+        <Link href="/calendar" className={styles.brandLink}>
+          <Image src="/logo.png" alt="Garden Workday Logo" width={55} height={61} className={styles.logo} />
+          <span className={styles.brandText}>GARDEN WORKDAY EVENTS</span>
+        </Link>
+      </div>
+
+      {/* RIGHT SIDE (your existing pill nav unchanged) */}
       <div className={styles.container}>
         {links.map((item) => {
           const active = isActivePath(pathname, item.href);
